% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_test_stat.R
\name{compute_test_stat}
\alias{compute_test_stat}
\title{Return statistics for covariates balance or placebo test.}
\usage{
compute_test_stat(
  df_data,
  int_dim_Z,
  weights = NULL,
  covs = NULL,
  kernel = "triangular",
  int_J = 3,
  bool_max_test = FALSE,
  bool_L2_std = FALSE,
  bool_joint = TRUE
)
}
\arguments{
\item{df_data}{Data.frame, data to evaluate.}

\item{int_dim_Z}{Integer, dimension of covariates.}

\item{weights}{Weights of observations that multiply the kernel function.}

\item{covs}{Data.frame or Vector, optional control covariates for rdrobust.}

\item{kernel}{is the kernel function used to construct the local-polynomial
estimator(s). Options are triangular (default option), epanechnikov and uniform.}

\item{int_J}{Integer, nearest neighbor for the variance estimation <= 3
default is 3.}

\item{bool_max_test}{Boolean, use max test instead of L2 test,
default is false.}

\item{bool_L2_std}{Boolean, for using standardized t stat for L2 test.}

\item{bool_joint}{Boolean, for returning joint test, instead of balance only.}
}
\description{
Return statistics for covariates balance or placebo test.
}
