% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatChiSquared.R
\name{computeStatChiSquared}
\alias{computeStatChiSquared}
\title{return chiSquared statistics for covariates balance}
\usage{
computeStatChiSquared(
  df.data,
  int.dimZ,
  bool.joint,
  int.J,
  bool.knownV = FALSE,
  real.covZ = NULL,
  fun.adjustBand = NULL,
  bool.chisqStd = FALSE
)
}
\arguments{
\item{df.data}{data.frame of data to evaluate}

\item{int.dimZ}{integer dimension of covariates}

\item{bool.joint}{Boolean TRUE if joint test, FALSE if covariate balance test only}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool.knownV}{boolean option for setting known V instead of estimated Vinv}

\item{real.covZ}{real value of correlation in V, nullified if bool.knownV is FALSE or in default}

\item{fun.adjustBand}{real-valued function which returns a fraction of reduced bandwidth}

\item{bool.chisqStd}{boolean option for using standardized t stat for L2 test}
}
\description{
return chiSquared statistics for covariates balance
}
