% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStatMax.R
\name{computeStatMax}
\alias{computeStatMax}
\title{return max statistics for covariates balance}
\usage{
computeStatMax(
  df.data,
  int.dimZ,
  bool.joint,
  int.J,
  bool.maxTestVinv,
  bool.knownV = FALSE,
  real.covZ = NULL,
  fun.adjustBand = NULL
)
}
\arguments{
\item{df.data}{data.frame of data to evaluate}

\item{int.dimZ}{integer dimension of covariates}

\item{bool.joint}{Boolean TRUE if joint test, FALSE if covariate balance test only}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool.maxTestVinv}{boolean option for using inv V for max test}

\item{bool.knownV}{boolean option for setting known V instead of estimated Vinv}

\item{real.covZ}{real value of correlation in V, nullified if bool.knownV is FALSE or in default}

\item{fun.adjustBand}{real-valued function which returns a fraction of reduced bandwidth}
}
\description{
return max statistics for covariates balance
}
