% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_option_simulation.R
\name{load_option_simulation}
\alias{load_option_simulation}
\title{Option loaders for simulations}
\usage{
load_option_simulation(list_option)
}
\arguments{
\item{list_option}{list}
}
\description{
Loading options for simulations tuned up for specified DGPs
}
