% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOptionsSimulation.R
\name{loadOptionsSimulation}
\alias{loadOptionsSimulation}
\title{Option loaders for simulations}
\usage{
loadOptionsSimulation(list.options)
}
\arguments{
\item{list.options}{list}
}
\description{
Loading options for simulations tuned up for specified DGPs
}
