% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtest.R
\name{rdtest}
\alias{rdtest}
\title{Main function of the joint RD tests}
\usage{
rdtest(
  df_Z,
  vec_X,
  bool_balance_only,
  int_J = 3,
  real_cutoff = 0,
  bool_max_test = TRUE
)
}
\arguments{
\item{df_Z}{A data frame of pre-determined covariates.}

\item{vec_X}{A vector of the running variable X.}

\item{bool_balance_only}{A boolean option, if TRUE, then balance test of z
runs}

\item{int_J}{An integer of nearest neighbor for the variance estimation, the
default value is 3 and the value should be <= 3.}

\item{real_cutoff}{A scalar real of cutoff value, default = 0.}

\item{bool_max_test}{A boolean option, if True, then max test runs instead of
L2 test.}
}
\value{
A list of list_result containing test results.
}
\description{
rdtest is ...
}
