% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_random_stat.R
\name{return_random_stat}
\alias{return_random_stat}
\title{return_random_stat}
\usage{
return_random_stat(int_num_simul, int_dim, mat_cor)
}
\arguments{
\item{int_num_simul}{Integer, number of simulation draws.}

\item{int_dim}{Integer, dimension to generate the draws.}

\item{mat_cor}{Matrix, correlation matrix of the target random variable.}
}
\description{
an internal function for critical value computation
}
