% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_result_balance.R
\name{return_result_balance}
\alias{return_result_balance}
\title{return covariates balance test results}
\usage{
return_result_balance(df_data, int_dimZ, int_J = 3, bool_maxTest = FALSE)
}
\arguments{
\item{df.data}{data.frame of data to evaluate}

\item{int.dimZ}{integer dimension of covariates}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool.maxTest}{boolean option to use max test rather than chi-squared test}
}
\description{
return covariates balance test results
}
