% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsBalanceTest.R
\name{returnResultsBalanceTest}
\alias{returnResultsBalanceTest}
\title{return covariates balance test results}
\usage{
returnResultsBalanceTest(df.data, int.dimZ, int.J = 3, bool.maxTest = FALSE)
}
\arguments{
\item{df.data}{data.frame of data to evaluate}

\item{int.dimZ}{integer dimension of covariates}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool.maxTest}{boolean option to use max test rather than chi-squared test}
}
\description{
return covariates balance test results
}
