% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsChiIIndependence.R
\name{returnResultsChiIndependence}
\alias{returnResultsChiIndependence}
\title{return chi square test results of indepdendence}
\usage{
returnResultsChiIndependence(df.check, int.numPartition)
}
\arguments{
\item{df.check}{data.frame of data to evaluate}

\item{int.numPartition}{integer dimension of partitions}
}
\description{
return chi square test results of indepdendence
}
