% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsJointTestAdjustedSize.R
\name{returnResultsJointTestAdjustedSize}
\alias{returnResultsJointTestAdjustedSize}
\title{return joint test result}
\usage{
returnResultsJointTestAdjustedSize(
  df.data,
  int.dimZ,
  int.J = 3,
  bool.maxTest = FALSE,
  bool.maxTestVinv = FALSE,
  bool.knownV = FALSE,
  real.covZ = NULL,
  fun.adjustBand = NULL,
  vec.critRange = vec.critRange
)
}
\arguments{
\item{df.data}{data.frame of data to evaluate}

\item{int.dimZ}{integer dimension of covariates}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool.maxTest}{boolean option to use max test rather than chi-squared test}

\item{bool.maxTestVinv}{boolean option for using inv V for max test}

\item{bool.knownV}{boolean option for setting known V instead of estimated Vinv}

\item{real.covZ}{real value of correlation in V, nullified if bool.knownV is FALSE or in default}

\item{fun.adjustBand}{real-valued function which returns a fraction of reduced bandwidth}

\item{vec.critRange}{vector of critical values (in nominal size) to find the size adjustable critical value}
}
\description{
return joint test result
}
