% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsJointTestVariableCritical.R
\name{returnResultsJointTestVariableCritical}
\alias{returnResultsJointTestVariableCritical}
\title{return joint test result}
\usage{
returnResultsJointTestVariableCritical(
  df.data,
  int.dimZ,
  int.J = 3,
  bool.maxTest = FALSE,
  bool.maxTestVinv = FALSE,
  bool.knownV = FALSE,
  real.covZ = NULL,
  fun.adjustBand = NULL,
  real.criticalAlt = 0.95
)
}
\arguments{
\item{df.data}{data.frame of data to evaluate}

\item{int.dimZ}{integer dimension of covariates}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool.maxTest}{boolean option to use max test rather than chi-squared test}

\item{bool.maxTestVinv}{boolean option for using inv V for max test}

\item{bool.knownV}{boolean option for setting known V instead of estimated Vinv}

\item{real.covZ}{real value of correlation in V, nullified if bool.knownV is FALSE or in default}

\item{fun.adjustBand}{real-valued function which returns a fraction of reduced bandwidth}

\item{real.criticalAlt}{real-valued critical value (in nominal coverage) to adjust the size of the joint test}
}
\description{
return joint test result
}
