% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsMCCovariateBalance.R,
%   R/returnResultsMCCovariateBalanceCustom.R
\name{returnResultsMCCovariateBalance}
\alias{returnResultsMCCovariateBalance}
\title{MC covarage / power check using simulated data}
\usage{
returnResultsMCCovariateBalance(
  jump,
  dim,
  n,
  cov.z,
  frac.jump = 0,
  bool.mutePrint = FALSE,
  bool.maxTest = FALSE
)

returnResultsMCCovariateBalance(
  jump,
  dim,
  n,
  cov.z,
  frac.jump = 0,
  bool.mutePrint = FALSE,
  bool.maxTest = FALSE
)
}
\arguments{
\item{jump}{z jump size}

\item{dim}{number of z variables}

\item{n}{sample size}

\item{frac.jump}{fraction of z with jumps specified as jump. 1 if only the last one, 0 if none, 0 < frac < 1 if fraction of vector, default: null. 0}

\item{bool.mutePrint}{Boolean option for muting outputs}

\item{bool.maxTest}{boolean option for using the max test instead of chisqaured test, default is false}

\item{cov}{correlation of Z variables}
}
\description{
Example code for simulation excercise

Example code for simulation excercise
}
