% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsMCJointTest.R
\name{returnResultsMCJointTest}
\alias{returnResultsMCJointTest}
\title{MC joint test of coverage / power check using simulated data}
\usage{
returnResultsMCJointTest(
  int_ns = 300,
  x_jump,
  jump,
  a_2,
  dim,
  n,
  cov_z,
  frac_jump = "Half",
  bool_mutePrint = FALSE,
  int_J = 3,
  bool_maxTest = FALSE,
  bool_maxTestVinv = FALSE,
  bool_knownV = FALSE,
  fun_adjustBand = NULL,
  bool_chisqStd = FALSE
)
}
\arguments{
\item{int_ns}{Interger, number of simulations to run
default is 300.}

\item{x_jump}{Real, x jump size.}

\item{jump}{Real, z jump size.}

\item{a_2}{Real, gap in heteroskedasticity.}

\item{dim}{Integer, number of z variables.}

\item{n}{Integer, sample size.}

\item{frac_jump}{String, either "None", "Half", "All", "Last".}

\item{bool_mutePrint}{Boolean, option for muting outputs.
default is FALSE}

\item{int_J}{Integer, nearest neighbor for the variance estimation <= 3
default is 3.}

\item{bool_maxTest}{Boolean, use max test instead of L2 test,
default is false.}

\item{bool_maxTestVinv}{Boolean, option for using inverse V for max test.}

\item{bool_knownV}{Boolean, option for setting V as known.}

\item{bool_chisqStd}{Boolean, for using standardized t stat for L2 test.}

\item{cov}{Real, correlation of Z variables.}

\item{fun.adjustBand}{Function, returns a fraction of reduced bandwidth.}
}
\description{
Example code for simulation excercise
}
