% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsMCJointTestVariableCritical.R
\name{returnResultsMCJointTestVariableCritical}
\alias{returnResultsMCJointTestVariableCritical}
\title{MC joint test of coverage / power check using simulated data}
\usage{
returnResultsMCJointTestVariableCritical(
  int.ns = 300,
  x.jump,
  jump,
  a.2,
  dim,
  n,
  cov.z,
  frac.jump = "Half",
  bool.mutePrint = FALSE,
  int.J = 3,
  bool.maxTest = FALSE,
  bool.maxTestVinv = FALSE,
  bool.knownV = FALSE,
  fun.adjustBand = NULL,
  real.criticalAlt = 0.95
)
}
\arguments{
\item{int.ns}{number of simulations to run}

\item{x.jump}{x jump size}

\item{jump}{z jump size}

\item{a.2}{gap in heteroskedasticity}

\item{dim}{number of z variables}

\item{n}{sample size}

\item{frac.jump}{Either None, Half, All, Last}

\item{bool.mutePrint}{boolean option for muting outputs}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool.maxTest}{boolean option for using the max test instead of chisqaured test, default is false}

\item{bool.maxTestVinv}{boolean option for using inverse V for max test}

\item{bool.knownV}{boolean option for setting V as known (instead of inverse V)}

\item{fun.adjustBand}{real-valued function which returns a fraction of reduced bandwidth}

\item{real.criticalAlt}{real-valued critical value (in nominal coverage) to adjust the size of the joint test}

\item{cov}{correlation of Z variables}
}
\description{
Example code for simulation excercise
}
