% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnResultsMCOrthogonalStats.R
\name{returnResultsMCOrthogonalStats}
\alias{returnResultsMCOrthogonalStats}
\title{MC orthogonality check}
\usage{
returnResultsMCOrthogonalStats(jump, a.2, n, bool.mutePrint = FALSE)
}
\arguments{
\item{jump}{x jump size}

\item{a.2}{heteroskedasticity added at cutoff}

\item{n}{sample size}

\item{bool.mutePrint}{boolean option to mute all the outputs}
}
\description{
MC example of simulation study for orthogonality
}
