% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_DGP.R
\name{simulate_DGP}
\alias{simulate_DGP}
\title{Simulating DGP}
\usage{
simulate_DGP(list_option)
}
\arguments{
\item{list_option}{List, list of options that specify the DGP.}
}
\description{
Simulation code for DGP
}
