% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateCustomDGPs.R
\name{simulateCustomDGPs}
\alias{simulateCustomDGPs}
\title{Simulating customized DGP using utilRDMC}
\usage{
simulateCustomDGPs(int.typeDGP, int.dim, int.sampleSize, real.TE, real.corr)
}
\arguments{
\item{int.typeDGP}{an integer representing the type of custom DGP}

\item{int.dim}{an integer of covariates dimensions}

\item{int.sampleSize}{an integer of sample size}

\item{real.TE}{real value of the discontinuity gap}

\item{real.corr}{real value of correlation coefficient across dep. vars}
}
\description{
Simulating customized DGP using utilRDMC
}
