% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtest.R
\name{summary.rdtest}
\alias{summary.rdtest}
\title{Summary function of the main caller.}
\usage{
\method{summary}{rdtest}(object, ...)
}
\arguments{
\item{object}{Object rdtest, result returned from rdtest function.}

\item{...}{Other options.}
}
\description{
Summary function of the main caller.
}
