% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtest.R
\name{rdtest}
\alias{rdtest}
\title{Main Test Function}
\usage{
rdtest(Z, X, bool_balance, int_J = 3, real_cutoff = 0, bool_max_test = TRUE)
}
\arguments{
\item{Z}{data frame of pre-determined covariates}

\item{X}{vector of the running variable X}

\item{bool_balance}{Boolean TRUE if joint balance test of z should be run}

\item{int_J}{integer of nearest neighbor for the variance estimation <= 3}

\item{real_cutoff}{scalar real of cutoff value, default = 0}

\item{bool_max_test}{Boolean,}
}
\description{
Main Test Function
}
