% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtest.R
\name{rdtest}
\alias{rdtest}
\title{rdtest: A joint diagnosis for regression discontinuity designs.}
\usage{
rdtest(
  Z,
  vec_X,
  bool_joint = TRUE,
  int_J = 3,
  real_cutoff = 0,
  bool_max_test = FALSE,
  bool_L2_std = TRUE
)
}
\arguments{
\item{Z}{Data.frame or Vector, pre-determined covariates.}

\item{vec_X}{Vector, running variable X.}

\item{bool_joint}{Boolean, TRUE if joint test instead of balance test only,
default is TRUE.}

\item{int_J}{Integer, the number or nearest neighbor for the variance
estimation, default is 3.}

\item{real_cutoff}{Real scalar, cutoff value, default = 0.}

\item{bool_max_test}{Boolean, use of max test, instead of Wald tests, default
is FALSE.}

\item{bool_L2_std}{Boolean, use of standardized Wald test, default is TRUE.}
}
\description{
The rdtest package offers a joint test of commonly used diagnostic tests for
regression discontinuity designs.
}
\examples{
# Prepare a mock dataset
library(rdtest)
set.seed(1)
N <- 1000
Z <- data.frame(
 var1 = rnorm(N),
 var2 = rnorm(N),
 var3 = rnorm(N),
 var4 = rnorm(N),
 var5 = rnorm(N))
 vec_X <- rnorm(N)
# Returning result with sWald test statistic
res_rd <- rdtest::rdtest(Z = Z, vec_X = vec_X)
summary(res_rd)
# Returning result with max test statistic
res_rd <- rdtest::rdtest(Z = Z, vec_X = vec_X, bool_max_test = TRUE)
summary(res_rd)
}
