% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtest.R
\name{rdtest}
\alias{rdtest}
\title{Main Test Function}
\usage{
rdtest(Z, X, bool.BalanceOnly, int.J = 3, real.cutoff = 0, bool.maxTest = TRUE)
}
\arguments{
\item{Z}{data frame of pre-determined covariates}

\item{X}{vector of the running variable X}

\item{bool.BalanceOnly}{Boolean TRUE if joint balance test of z should be run}

\item{int.J}{integer of nearest neighbor for the variance estimation <= 3}

\item{real.cutoff}{scalar real of cutoff value, default = 0}
}
\description{
Main Test Function
}
