% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtest.R
\name{rdtest}
\alias{rdtest}
\title{Main Test Function}
\usage{
rdtest(
  Z,
  vec_X,
  bool_joint = TRUE,
  int_J = 3,
  real_cutoff = 0,
  bool_max_test = FALSE,
  bool_max_test_V_inv = FALSE,
  bool_L2_std = TRUE
)
}
\arguments{
\item{Z}{Data.frame or Vector, pre-determined covariates}

\item{vec_X}{Vector, running variable X}

\item{bool_joint}{Boolean, TRUE if joint test instead of balance test only.}

\item{int_J}{integer of nearest neighbor for the variance estimation <= 3}

\item{real_cutoff}{scalar real of cutoff value, default = 0}

\item{bool_max_test}{Boolean,}

\item{bool_max_test_V_inv}{Boolean, option for using inverse V for max test.}

\item{bool_L2_std}{Boolean, for using standardized t stat for L2 test.}
}
\description{
Main Test Function
}
