% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_result_balance.R
\name{return_result_balance}
\alias{return_result_balance}
\title{Return result of joint covariates balance test}
\usage{
return_result_balance(
  df_data,
  int_dim_Z,
  int_J = 3,
  bool_max_test = FALSE,
  bool_max_test_inv_V = FALSE,
  bool_L2_std = FALSE
)
}
\arguments{
\item{df_data}{A data.frame of data to evaluate.}

\item{int_dim_Z}{An integer dimension of covariates.}

\item{int_J}{An integer of nearest neighbor for the variance estimation <= 3.}

\item{bool_max_test}{A boolean option to return max test rather than L2 test.}

\item{bool_max_test_inv_V}{A boolean option for using inv V for max test.}

\item{bool_L2_std}{A boolean option for using standardized t stat to conduct
L2 test.}
}
\description{
Return result of joint covariates balance test
}
