% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_result_joint.R
\name{return_result_joint}
\alias{return_result_joint}
\title{return joint test result}
\usage{
return_result_joint(
  df_data,
  int_dimZ,
  int_J = 3,
  bool_maxTest = FALSE,
  bool_maxTestVinv = FALSE,
  bool_knownV = FALSE,
  real_covZ = NULL,
  fun_adjustBand = NULL,
  bool_chisqStd = FALSE
)
}
\arguments{
\item{df_data}{data.frame of data to evaluate}

\item{int_dimZ}{integer dimension of covariates}

\item{int_J}{integer of nearest neighbor for the variance estimation <= 3}

\item{bool_maxTest}{boolean option to use max test rather than chi-squared test}

\item{bool_maxTestVinv}{boolean option for using inv V for max test}

\item{bool_knownV}{boolean option for setting known V instead of estimated Vinv}

\item{real_covZ}{real value of correlation in V, nullified if bool.knownV is FALSE or in default}

\item{fun_adjustBand}{real-valued function which returns a fraction of reduced bandwidth}

\item{bool_chisqStd}{boolean option for using standardized t stat to conduct L2 test}
}
\description{
return joint test result
}
