% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_result_joint.R
\name{return_result_joint}
\alias{return_result_joint}
\title{return joint test result}
\usage{
return_result_joint(
  df_data,
  int_dim_Z,
  int_J = 3,
  bool_max_test = FALSE,
  bool_L2_std = TRUE,
  bool_joint = TRUE,
  bool_equivalence = FALSE,
  real_epsilon = NULL
)
}
\arguments{
\item{df_data}{Data.frame, data to evaluate.}

\item{int_dim_Z}{Integer, dimension of covariates.}

\item{int_J}{Integer, nearest neighbor for the variance estimation <= 3
default is 3.}

\item{bool_max_test}{Boolean, use max test instead of L2 test,
default is false.}

\item{bool_L2_std}{Boolean, for using standardized t stat for L2 test.}

\item{bool_joint}{Boolean, for returning joint test, instead of balance only.}

\item{bool_equivalence}{Boolean, for returning equivalence test.}

\item{real_epsilon}{Numeric value, for equivalence test. Default NULL.}
}
\description{
return joint test result
}
